Создает контроллер для отрисовки доски с задачами
